requirements.txt
anyio==3.6.2
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==3.4.8
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.92.0
fastapi-jwt-auth==0.5.0
h11==0.14.0
idna==3.4
motor==3.1.1
numpy==1.24.2
pandas==1.5.3
passlib==1.7.4
pycparser==2.21
pydantic==1.10.5
PyJWT==1.7.1
pymongo==4.3.3
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0